requirements.txt
# Web framework & templates
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.20

# AI - Google Gemini
google-genai==1.0.0

# Supabase
supabase==2.10.0

# Config & utilities
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx>=0.26,<0.28
Pillow==11.1.0
aiofiles==24.1.0
